+++ Editor/Core/ResourceRegistry.cs
@@ -123,7 +123,7 @@ namespace UnityMcp {
                 name = t.name,
                 activeSelf = t.gameObject.activeSelf,
                 activeInHierarchy = t.gameObject.activeInHierarchy,
-                instanceId = EntityIdCompat.GetId(t.gameObject),
+                instanceId = EntityIdCompat.GetIdString(t.gameObject),
                 childCount = t.childCount
             });